Feedback from bitfocus addressed
You are missing a HELP.md file in the /companion folder. This should have info for users to set up the module

I'm unclear why the /vendor folder is in the /companion directory?

The id in your manifest.json must match the module name exactly, so in this case it needs to be "meld-studio"

Your manifest.json "repository" and "bugs" URL should replace yourname with bitfocus

In manifest.json the "runtime"> "apiVersion" should be left at 0.0.0

In package.json, version number should match the tag, without the v. So if tag is v1.0.0, version should be 1.0.0

The code does not appear to be formatted. Please install prettier as a devDependency, and then add "prettier": "@companion-module/tools/.prettierrc.json" to your package.json, then run prettier to format the code

You should add a .yarnrc.yml file like this and delete the .pnp.cjs and .pnp.loader.mjs, otherwise the build process will fail https://github.com/bitfocus/companion-module-template-js/blob/main/.yarnrc.yml